pull automation out of spell
